Use int* instead of size_t* to store sparsity (pretty much all subproblem solvers use int*)

+++ bindings/AMPL/AMPLModel.hpp
@@ -33,9 +33,9 @@ namespace uno {
       void evaluate_objective_gradient(const Vector<double>& x, Vector<double>& gradient) const override;
 
       // structures of Jacobian and Hessian
-      void compute_constraint_jacobian_sparsity(size_t* row_indices, size_t* column_indices, size_t solver_indexing,
+      void compute_constraint_jacobian_sparsity(int* row_indices, int* column_indices, int solver_indexing,
          MatrixOrder matrix_order) const override;
-      void compute_hessian_sparsity(size_t* row_indices, size_t* column_indices, size_t solver_indexing) const override;
+      void compute_hessian_sparsity(int* row_indices, int* column_indices, int solver_indexing) const override;
 
       // numerical evaluations of Jacobian and Hessian
       void evaluate_constraint_jacobian(const Vector<double>& x, double* jacobian_values) const override;
