+++ uno/ingredients/constraint_relaxation_strategies/l1RelaxedProblem.cpp
@@ -103,40 +103,40 @@ namespace uno {
       return number_nonzeros;
    }
 
-   void l1RelaxedProblem::compute_constraint_jacobian_sparsity(size_t* row_indices, size_t* column_indices, size_t solver_indexing,
+   void l1RelaxedProblem::compute_constraint_jacobian_sparsity(int* row_indices, int* column_indices, int solver_indexing,
          MatrixOrder matrix_order) const {
       this->model.compute_constraint_jacobian_sparsity(row_indices, column_indices, solver_indexing, matrix_order);
 
       // add the contribution of the elastic variables
-      size_t elastic_index = this->model.number_variables;
+      int elastic_index = static_cast<int>(this->model.number_variables);
       size_t current_index = this->model.number_jacobian_nonzeros();
       for (size_t inequality_index: this->model.get_inequality_constraints()) {
-         row_indices[current_index] = inequality_index + solver_indexing;
+         row_indices[current_index] = static_cast<int>(inequality_index) + solver_indexing;
          column_indices[current_index] = elastic_index + solver_indexing;
          ++elastic_index;
          ++current_index;
       }
       for (size_t equality_index: this->model.get_equality_constraints()) {
-         row_indices[current_index] = equality_index + solver_indexing;
+         row_indices[current_index] = static_cast<int>(equality_index) + solver_indexing;
          column_indices[current_index] = elastic_index + solver_indexing;
          ++current_index;
-         row_indices[current_index] = equality_index + solver_indexing;
+         row_indices[current_index] = static_cast<int>(equality_index) + solver_indexing;
          column_indices[current_index] = elastic_index + 1 + solver_indexing;
          elastic_index += 2;
          ++current_index;
       }
    }
 
-   void l1RelaxedProblem::compute_hessian_sparsity(const HessianModel& hessian_model, size_t* row_indices,
-         size_t* column_indices, size_t solver_indexing) const {
+   void l1RelaxedProblem::compute_hessian_sparsity(const HessianModel& hessian_model, int* row_indices,
+         int* column_indices, int solver_indexing) const {
       hessian_model.compute_sparsity(this->model, row_indices, column_indices, solver_indexing);
 
       // diagonal proximal contribution
       if (this->proximal_center != nullptr && this->proximal_coefficient != 0.) {
          size_t current_index = hessian_model.number_nonzeros(this->model);
          for (size_t variable_index: Range(this->model.number_variables)) {
-            row_indices[current_index] = variable_index + solver_indexing;
-            column_indices[current_index] = variable_index + solver_indexing;
+            row_indices[current_index] = static_cast<int>(variable_index) + solver_indexing;
+            column_indices[current_index] = static_cast<int>(variable_index) + solver_indexing;
             ++current_index;
          }
       }
