+++ uno/model/BoundRelaxedModel.hpp
@@ -27,12 +27,12 @@ namespace uno {
       }
 
       // sparsity patterns of Jacobian and Hessian
-      void compute_constraint_jacobian_sparsity(size_t* row_indices, size_t* column_indices, size_t solver_indexing,
+      void compute_constraint_jacobian_sparsity(int* row_indices, int* column_indices, int solver_indexing,
             MatrixOrder matrix_order) const override {
          this->model->compute_constraint_jacobian_sparsity(row_indices, column_indices, solver_indexing, matrix_order);
       }
 
-      void compute_hessian_sparsity(size_t* row_indices, size_t* column_indices, size_t solver_indexing) const override {
+      void compute_hessian_sparsity(int* row_indices, int* column_indices, int solver_indexing) const override {
          this->model->compute_hessian_sparsity(row_indices, column_indices, solver_indexing);
       }
